+++ src/router/algorithm/service/get_algorithm_count.ts
@@ -1,0 +1,12 @@
+import { getCustomRepository } from "typeorm";
+import { AlgorithmRepository } from "../../../repository";
+import { createRes } from "../../../util/http";
+
+const getAlgorithmCountAtAll = async (connectionName: string) =>
+  createRes({
+    data: await getCustomRepository(
+      AlgorithmRepository,
+      connectionName,
+    ).getAlgorithmCountAtAll(),
+  });
+export default getAlgorithmCountAtAll;
