+++ src/router/algorithm/service/get_web_rule.ts
@@ -1,0 +1,10 @@
+import { ruleForWeb } from "../../../config";
+import { createRes } from "../../../util/http";
+
+const getAlgorithmRulesForWeb = () =>
+  createRes({
+    data: {
+      content: ruleForWeb,
+    },
+  });
+export default getAlgorithmRulesForWeb;
