+++ src/router/algorithm/service/index.ts
@@ -1,0 +1,21 @@
+import deleteAlgorithm from "./delete";
+import getAlgorithmByIdx from "./get_alone_algorithm";
+import getAlgorithmCountAtAll from "./get_algorithm_count";
+import getAlgorithmLists from "./get_algorithm_list";
+import getAlgorithmRules from "./get_rule";
+import getAlgorithmRulesForWeb from "./get_web_rule";
+import modifyAlgorithmContent from "./modify";
+import setAlgorithmStatus from "./set_status";
+import writeAlgorithm from "./write";
+
+export {
+  deleteAlgorithm,
+  getAlgorithmByIdx,
+  getAlgorithmCountAtAll,
+  getAlgorithmLists,
+  getAlgorithmRules,
+  getAlgorithmRulesForWeb,
+  modifyAlgorithmContent,
+  setAlgorithmStatus,
+  writeAlgorithm,
+};
