+++ src/router/auth/index.ts
@@ -1,43 +1,41 @@
-import { Context } from "aws-lambda";
 import { APIGatewayEventIncludeConnectionName } from "../../DTO/http.dto";
 import { QuestionDTO } from "../../DTO/question.dto";
 import { AuthMiddleware, DBMiddleware } from "../../middleware";
 import { HttpErrorException } from "../../middleware/error";
 import { getBody } from "../../util/req";
-
-import { AuthService } from "./auth.service";
+import {
+  signUp,
+  login,
+  mailSend,
+  mailAuth,
+  addVerifyQuestion,
+  getTokenByRefreshToken,
+  getVerifyQuestion,
+} from "./service";
 
 export class AuthRouter {
   @HttpErrorException
   @DBMiddleware.connectTypeOrm
-  static async signUp(
-    event: APIGatewayEventIncludeConnectionName
-  ) {
-    return AuthService.signUp(event);
+  static async signUp(event: APIGatewayEventIncludeConnectionName) {
+    return signUp(event);
   }
 
   @HttpErrorException
   @DBMiddleware.connectTypeOrm
-  static async login(
-    event: APIGatewayEventIncludeConnectionName,
-  ) {
-    return AuthService.login(event);
+  static async login(event: APIGatewayEventIncludeConnectionName) {
+    return login(event);
   }
 
   @HttpErrorException
   @DBMiddleware.connectTypeOrm
-  static async mailSend(
-    event: APIGatewayEventIncludeConnectionName
-  ) {
-    return AuthService.mailSend(event);
+  static async mailSend(event: APIGatewayEventIncludeConnectionName) {
+    return mailSend(event);
   }
 
   @HttpErrorException
   @DBMiddleware.connectTypeOrm
-  static async mailAuth(
-    event: APIGatewayEventIncludeConnectionName
-  ) {
-    return AuthService.mailAuth(event);
+  static async mailAuth(event: APIGatewayEventIncludeConnectionName) {
+    return mailAuth(event);
   }
 
   @HttpErrorException
@@ -46,14 +44,14 @@ export class AuthRouter {
   static async getVerifyQuestion({
     connectionName,
   }: APIGatewayEventIncludeConnectionName) {
-    return AuthService.getVerifyQuestion(connectionName);
+    return getVerifyQuestion(connectionName);
   }
 
   @HttpErrorException
   @AuthMiddleware.authAdminPassword
   @DBMiddleware.connectTypeOrm
   static async addVerifyQuestion(event: APIGatewayEventIncludeConnectionName) {
-    return AuthService.addVerifyQuestion(
+    return addVerifyQuestion(
       getBody<QuestionDTO>(event.body),
       event.connectionName,
     );
@@ -66,8 +64,8 @@ export class AuthRouter {
     headers,
     connectionName,
   }: APIGatewayEventIncludeConnectionName) {
-    return AuthService.getTokenByRefreshToken(
-      headers.Authorization || headers.authorization,
+    return getTokenByRefreshToken(
+      headers.Authorization || headers.authorization || "",
       connectionName,
     );
   }
