+++ src/router/auth/service/index.ts
@@ -1,0 +1,17 @@
+import addVerifyQuestion from "./add_verify";
+import getTokenByRefreshToken from "./get_token";
+import getVerifyQuestion from "./get_verify";
+import login from "./login";
+import mailAuth from "./mail_auth";
+import mailSend from "./mail_send";
+import signUp from "./regist";
+
+export {
+  addVerifyQuestion,
+  getTokenByRefreshToken,
+  getVerifyQuestion,
+  login,
+  mailAuth,
+  mailSend,
+  signUp,
+};
