+++ src/router/comment/service/index.ts
@@ -1,0 +1,4 @@
+import addComment from "./add";
+import deleteComment from "./delete";
+
+export { addComment, deleteComment };
