+++ src/router/emoji/service/index.ts
@@ -1,0 +1,4 @@
+import addLeaf from "./add";
+import removeLeaf from "./remove";
+
+export { addLeaf, removeLeaf };
